Switch to using the system command line options.
Added the copyright & license notices in the command line options.
Put arguments in two groups: COMMANDS & OPTIONS.
Added a catch so we can have the auto-execution of the system commands.
Removed the handling of the help & version since that's part of the system commands.
Fixed an access to the "reset" flag, it should have used the f_reset field.
Fixed a message which had a hard coded size instead of IFNAMSIZ.
Added the mk script so I can quickly recompile and install this project.
Removed the tokenize_string.h since we're not using it (and we'd want to use the snapdev/... version instead.)