Changed docs about order.
For us to properly order the rules we can have sections (group of rules) and
a before and an after list of rules (to sort the sections, there is also
a before and an after for the sections).

The sort is super important since the final list of rules must be sorted
as expected in the final iptables list.