Worked on the new packages.
This includes a cmake directory.
Moved the ipload files in a sub-directory we can install at once.
Added the .install and other debian files for ipwall and iplock-dev.
Install the services, especially, the ipload.service is a separate file not
under the debian directory since the name of the package is iplock...
The ipwall package depends on iplock so ipload can just reside in iplock.
Applied another path against the iplock source files to make it better.
I added a command line option to define the cache folder (for example).
And I modernized a few more things and noticed yet nore commented out #include
and code which I removed.