Added default rules in case of failures.
The code will now first install the basic rules.
Try to load the user defined configuration files.
If no configuration is loaded, it falls back on a default firewall.
If the configuration is loaded, we try to install it in iptables. If that
fails, we try to install the default firewall.

Now I have to do the actual IP rules configuration to iptables rules.