Here is a version that works.
This version is able to verify, show, load the firewall rules.
The order of rules and section was fixed by using a dependency tree.
The rules were reorganized properly (in the correct section as first
intended--have the `--show` and `--show-dependencies` options really
helped).
I added rule samples for FTP, HTTP, SMTP.
Fixed several variables (references, names, content).
Added support for sets and descriptions and comments in the output...