Added rules for our IPv6 firewall.
Verify addresses as valid IPv4 or IPv6 addresses.
Added proper support for icmp and icmpv6.
Moved the "unwanted" rules from the "early_content" section to the "content" folder.
Added "::1" as the IPv6 localhost address.
Made a single basic.rules file (we can use the --ipv4 or --ipv6 flags to create rules that are target to one of the specific system).