Moved iptables logrotate conf to iplock.
Removed some more fail2ban setup remains.
Changed icmp in IPv4 to no protocol when it comes down to "lo" in the OUTPUT chain.
Added target synonyms: ALLOW for ACCEPT, and DENY and BLACKHOLE for DROP.
Cleaned up the manual pages. There is still a lot of work to do in there...
Also fixed inline comments and added a comment to the basic.rules file.
(because ip[6]tables-restore finally support comments and empty lines!)