Write a knock-knock tool.
In order to better protect the SSH port, one can use a port knocker.
As long as the port knocker is managed by the firewall (through the "recent"
extension) and the port "opens" (is accessible, really) only if you get the
knocking in the correct order.

Also the order of the ESTABLISHED,RELATED rule was wrong. Testing to see that
it's now at the right place.