Started clean ups of the rules.
Added a script to compute the number of possibilities when using 2 or more
ports knocking. The math takes in account the number of ports that you use
for the knocking sequence, the number of ports in the destination, the fact
that you want to use each port only once, and that the destination port
cannot be included in the knocking sequence list of ports.