Many enhancements to the default configuration.
Especially, added support for concatenation & "optional" arguments (can be
set prior to loading this file).

Added support for conditional chains.

Added support for conntrack options (I use those with Docker).

Added a README.md for the ipload folders. It has some documentation which is
a complement to the manual page.