Further corrections/tweaks to ipload definitions.
I moved more of the variables to the one file where it gets used and document
them.

Fixed the source/destination for some ports of services (it was inverted).

Changed the "required" section with a "verify" section. That way we can
defined a "type" of verification. For now we support "defined" (i.e. the
variable is set with whatever value) and "required" (i.e. the value cannot
be empty).