Fixed the REJECT parameter.
The parameter may apply to IPv4 or IPv6 or both with different names. Therefore
I needed to have a table that would allow me to fix the name correctly in all
cases. I moves the strings as standalone and then built a table from those.
The parser defines both strings as required and the output code selects
IPv4 for the IPv4 rule and IPv6 for the IPv6 rule.