Made further updates to the default firewall configuration files.
Updated the LICENSE.txt and debian/copyright. It's GPL v3 now.
Fixed a CMakeLists.txt which still had the old license.
Fixed the add with a load. The `restore` command is much faster to load ipsets.
Fixed order in which
Added descriptions for documentation purposes.
Added a RETURN for admins in the unwanted chain. This means admins do not get blocked (at least those with a static IP).
Updated/enhanced the documentation (man pages, README, etc.)
Fixed the ipload command line definitions.
Added support for an ipload.conf file (useful for --rules and --ip-lists.
Fixed the multiport usage (if both, sources and destinations are defined, then -m multiport does not work).
Added support for the INVALID state.
Added support for the TCP MSS option.
Removed the "name" support in the section, we cannot replace the name specified in the '[...]'.