Fixed how ip-list is split (':', not ',').
Place the drop-list inside the ip-list folder. It could be other types of
IPs so we want to keep them in a generically named folder.

Added --load when calling the ipload tool.

Added triggers so when you place new ipload files when instaling another
package, it automatically run ipload for you.