Fixed the chown of /var/log/iptables.
Test the maximum of the recent hitcount. It is 255.
Changed the synflood to 250 hits in 20 seconds... but even that blocks my IP
address after just a little time. So I have to revisit and see whether these
rules are correct and/or whether my servers communicate way more than I ever
imagined...