Use dh_installsystemd to start ipload.
The systemctl has many extensions one of which is to properly start & stop
services inside .postinst/prerm scripts. This version takes advantage of
the possible override of systemd install scripts. This allows us to use
system defined scripts to start & stop our ipload service.

I also first initialize the ipload.conf variable file to tweak the firewall
properly for the current system before setting it up.

As a result, the ipload.service becomes iplock.ipload.service. It is otherwise
the exact same file and it will still be installed as just ipload.service.