Changed note about nftables.
The newer kernels all support the bfs filtering system which, apparently,
is a replacement of both: iptables and nftables. This is in part because
it is much faster to check packets with bfs than either iptables or nftables.
So I updated our TODO and we can test once we have time to work on this again.