Removed the callback from fluid.
The fluid object does not accept a callback anymore. That was my first way
to get a call whenever the STATUS command was received. Instead, we now
use the dispatcher with a CALLBACK type of dispatch. This is extensible
so any number of objects can receive the callback straight from the source.

Made "ipwall" a service name in the names.an file.
Make use of the communicatord "up" names.an value.