Updated the wait_on_firewall class.
The class would just wait for the ipwall status message
(IPWALL_CURRENT_STATUS), which is not the whole story. That is, if you
just need the firewall to be up, then checking that the ipload service
is active is sufficient.

Many other corrections went in (especially documentation wise where
I still had references to snapfirewall and snapcommunicator).