We mustn't depend on the default dependencies.
The default dependencies include a dependency on the network which needs to
be up and running when we want to run this service before that. Having such
a dependency creates a cycle and as a result our service does not always
start or starts late. With the extra parameter, I'm hoping I'm resolving
the whole issue.