Renamed communicatord to communicator.
Started small tweaks on wait_on_firewall. But I'll be transforming that
class in a plugin for the communicator daemon so we can use it in prinbee.
Also that way the test runs once for all the services instead of having
copies in each service and redo the work in each one of them.