Moving the IPWALL_... messages here.
The messages were defined in the communicator but now we have a plugin.
So instead, the IPWALL_... messages are defined along the ipwall service
and the IPLOCK_... messages are defined in the communicator.

All of those messages are understood by the wait_on_firewall plugin. The
IPLOCK_... can be used by individual services to ask the communicator whether
the firewall is properly setup.